Flatten namespace.
