+++ src/ImageSharp/Formats/Cur/CurConfigurationModule.cs
@@ -1,19 +1,21 @@
 // Copyright (c) Six Labors.
 // Licensed under the Six Labors Split License.
 
-namespace SixLabors.ImageSharp.Formats.Icon.Ico;
+using SixLabors.ImageSharp.Formats.Icon;
+
+namespace SixLabors.ImageSharp.Formats.Cur;
 
 /// <summary>
 /// Registers the image encoders, decoders and mime type detectors for the Ico format.
 /// </summary>
-public sealed class IcoConfigurationModule : IImageFormatConfigurationModule
+public sealed class CurConfigurationModule : IImageFormatConfigurationModule
 {
     /// <inheritdoc/>
     public void Configure(Configuration configuration)
     {
-        // TODO: IcoEncoder
-        // configuration.ImageFormatsManager.SetEncoder(IcoFormat.Instance, new IcoEncoder());
-        configuration.ImageFormatsManager.SetDecoder(IcoFormat.Instance, IcoDecoder.Instance);
+        // TODO: CurEncoder
+        // configuration.ImageFormatsManager.SetEncoder(CurFormat.Instance, new CurEncoder());
+        configuration.ImageFormatsManager.SetDecoder(CurFormat.Instance, CurDecoder.Instance);
         configuration.ImageFormatsManager.AddImageFormatDetector(new IconImageFormatDetector());
     }
 }
