+++ src/ImageSharp/Formats/Cur/CurConstants.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Six Labors.
 // Licensed under the Six Labors Split License.
 
-namespace SixLabors.ImageSharp.Formats.Icon.Cur;
+namespace SixLabors.ImageSharp.Formats.Cur;
 
 /// <summary>
 /// Defines constants relating to ICOs
